Optimise i18n template performance
Use an empty main.html and render dot template to support i18n, instead of reload entire page
