+++ src/app/assets/js/iSparta.js
@@ -1,5 +1,4 @@
 (function ($) {
-
 	var gui = require('nw.gui');
 	var win = gui.Window.get();
 	var os = require('os');
@@ -15,7 +14,6 @@
 				this.sep="\\";
 			}
 
-			window.iSparta.locale.init();
 			window.iSparta.ui.init();
 
 			this.checkVersion();
@@ -96,77 +94,6 @@
 			// 		});
 			// 	}
 			// });			
-		}
-	};
-
-		window.iSparta.locale = {
-		init: function() {
-			i18n.configure({
-				locales: ['en', 'zh-cn', 'zh-tw'],
-				directory: process.cwd() + '/app/locales',
-				defaultLocale: 'en',
-				fallbacks: {
-					'en-us': 'en',
-					'en-uk': 'en',
-					'zh_cn': 'zh-cn',
-					'zh-hans': 'zh-cn',
-					'zh-hans-cn': 'zh-cn',
-					'zh_tw': 'zh-tw',
-					'zh_hant': 'zh-tw',
-					'zh-hant-tw': 'zh-tw',
-				},
-				updateFiles: false
-			});
-			var currentLocale = this.getLocale();
-			if (!currentLocale) {
-				currentLocale = this.checkLocale();
-			}
-			this.setLocale(currentLocale);
-
-			var i18nTemplateSettings = {
-				evaluate:    /\[\[([\s\S]+?(\]?)+)\]\]/g,
-				interpolate: /\[\[=([\s\S]+?)\]\]/g,
-				encode:      /\[\[!([\s\S]+?)\]\]/g,
-				use:         /\[\[#([\s\S]+?)\]\]/g,
-				useParams:   /(^|[^\w$])def(?:\.|\[[\'\"])([\w$\.]+)(?:[\'\"]\])?\s*\:\s*([\w$\.]+|\"[^\"]+\"|\'[^\']+\'|\[[^\]]+\])/g,
-				define:      /\[\[##\s*([\w\.$]+)\s*(\:|=)([\s\S]+?)#\]\]/g,
-				defineParams:/^\s*([\w$]+):([\s\S]+)/,
-				conditional: /\[\[\?(\?)?\s*([\s\S]*?)\s*\]\]/g,
-				iterate:     /\[\[~\s*(?:\]\]|([\s\S]+?)\s*\:\s*([\w$]+)\s*(?:\:\s*([\w$]+))?\s*\]\])/g,
-				varname:	"i18n",
-				strip:		false,
-				append:		false,
-				selfcontained: false,
-				doNotSkipEncoded: false
-			};
-
-			var body = $(document.body).html();
-			var bodyTmpl = doT.template(body, i18nTemplateSettings);
-			var bodyHtml = bodyTmpl(i18n);
-			$(document.body).html(bodyHtml);
-		},
-		getLocale: function() {
-			return window.iSparta.localData.get('locale');
-		},
-		setLocale: function(locale) {
-			if (!locale) return;
-			i18n.setLocale(locale);
-			window.iSparta.localData.set('locale', i18n.getLocale());
-		},
-		checkLocale: function() {
-			var language = window.navigator.language;
-			var locale;
-			if (language) {
-				locale = language.toLowerCase();
-			} else {
-				locale = process.env.LANG.split('.')[0].toLowerCase();
-			}
-			return locale;
-		},
-		changeLocale: function(locale) {
-			if (!locale) return;
-			this.setLocale(locale);
-			win.reload();
 		}
 	};
