+++ src/app/assets/js/iSparta.pngloss.js
@@ -47,7 +47,7 @@
 			$.extend(this.options,options);
 			
 			options=this.options;
-			$currentLanguage.val(window.iSparta.locale.getLocale());
+			$currentLanguage.val(window.locale.getLocale());
 			
 			for(var i=0;i<options.savePath.length;i++){
 				if(options.savePath[i]=="parent"){
@@ -689,7 +689,7 @@
 			});
 			$currentLanguage.on('change', function() {
 				var locale=$(this).val();
-				window.iSparta.locale.changeLocale(locale);
+				window.locale.changeLocale(locale);
 			});
 		}
 	};
