+++ src/app/assets/js/iSparta.webp.js
@@ -68,7 +68,7 @@
             options = this.options;
             $config.val(options.config);
             $ratio.val(options.ratio);
-            $currentLanguage.val(window.iSparta.locale.getLocale());
+            $currentLanguage.val(window.locale.getLocale());
             
             for (var i=0; i<length; i++) {
                 if (options.savePath[i] == "parent") {
@@ -389,7 +389,7 @@
 
             $currentLanguage.on('change', function() {
                 var locale=$(this).val();
-                window.iSparta.locale.changeLocale(locale);
+                window.locale.changeLocale(locale);
             });
         },
