+++ src/app/assets/js/index.js
@@ -1,0 +1,3 @@
+$(document).ready(function(){
+	window.iSparta.init();
+});
