+++ src/app/locales/en.json
@@ -3,8 +3,8 @@
 	"PNG to APNG": "PNG to APNG",
 	"Images to WebP": "Images to WebP",
 	"APNG to Animated WebP": "APNG to Animated WebP",
-	"Images lossless compression": "Images lossless compression",
-	"PNG loss compression": "PNG loss compression",
+	"Images lossless": "Images lossless",
+	"PNG lossy": "PNG lossy",
 	"Loop count": "Loop count",
 	"forever": "forever",
 	"Frame rate": "Frame rate",
