+++ src/app/locales/zh-CN.json
@@ -3,8 +3,8 @@
 	"PNG to APNG": "PNG转APNG",
 	"Images to WebP": "图片转WebP",
 	"APNG to Animated WebP": "APNG转动态WebP",
-	"Images lossless compression": "图片无损压缩",
-	"PNG loss compression": "PNG有损压缩",
+	"Images lossless": "图片无损压缩",
+	"PNG lossy": "PNG有损压缩",
 	"Loop count": "循环次数",
 	"forever": "一直",
 	"Frame rate": "帧频率",
