+++ src/app/locales/zh-tw.json
@@ -3,8 +3,8 @@
 	"PNG to APNG": "PNG轉APNG",
 	"Images to WebP": "圖片轉WebP",
 	"APNG to Animated WebP": "APNG轉動態WebP",
-	"Images lossless compression": "圖片無損壓縮",
-	"PNG loss compression": "PNG有損壓縮",
+	"Images lossless": "圖片無損壓縮",
+	"PNG lossy": "PNG有損壓縮",
 	"Loop count": "循環次數",
 	"forever": "一直",
 	"Frame rate": "幀頻率",
